requirements.txt
Flask
flask_restful
requests
pyyaml
urllib3